Hand Positions tracking
Using OpenCV and mediapipe packages, tracking of 21 points in hand palm have been recorded in a file. Hands module in mediapipe is being used. Video of news reading with sign language is given as input to get the tracked positions. These positions can be used to apply movements to fingers of a robot/AI driven news reading